Split iOS Hyperliquid observer into subscription and event handler services
